Add helper functions for determining version of versioned types
